backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Limiter==2.8.1
Flask-Caching==2.0.1
python-dotenv==0.19.2
bcrypt==4.0.1
redis==4.3.4
SQLAlchemy==2.0.38
phonenumbers==8.13.22
requests==2.31.0
Flask-WTF==1.2.2
email-validator
Werkzeug==3.1.3
pytest~=8.3.4
WTForms~=3.2.1
alembic~=1.14.1
flask_cors
psycopg2-binary==2.9.9
pytest